some database migration fixes and seeder fixes

+++ database/migrations/2015_02_08_092923_diagnose_to_meta_fillable.php
@@ -47,8 +47,11 @@ class DiagnoseToMetaFillable extends Migration
 
         $context = Context::where('context', Context::OORZAAK_HULPBEHOEFTE)->first();
 
-        Value::where('context_id', $context->id)->delete();
-        $context->delete();
+        if($context)
+        {
+            Value::where('context_id', $context->id)->delete();
+            $context->delete();
+        }
 
         Context::where('context', Context::MANTELZORGER_RELATION)
             ->update(['context' => 'relation_mantelzorger_oudere']);
